requirements.txt
albumentations==1.3.1
einops==0.8.1
fairscale==0.4.13
h5py==3.12.1
lmdb==1.5.1
loguru==0.7.2
loralib==0.1.2
matplotlib==3.10.0
mc==0.0
numpy==1.25.2
omegaconf==2.3.0
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
peft==0.13.2
Pillow==11.1.0
pyiqa==0.1.13
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
skimage==0.0
tensorboard==2.18.0
thop==0.1.1.post2209072238
timm==1.0.14
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.6
transformers==4.37.2
wandb==0.19.7
xformers==0.0.23